+++ src/partials/header.html
@@ -1,30 +1,33 @@
 <header class="header">
-  <div class="header-container">
+  <div class="header-container container">
     <!-- Logo -->
     <div class="logo-header">
       <img src="img/logo/logotype.svg" alt="Logo" width="80" height="80">
       <span class="header-span">jefferson*</span>
     </div>
 
-    <!-- Кнопка відкриття бургер-меню (мобільні пристрої) -->
-    <button class="menu-toggle">
+    <!-- Мобільні пристрої -->
+    <button class="menu-toggle" type="button">
       <span class="burger-line"></span>
       <span class="burger-line"></span>
     </button>
 
     <!-- Навігація -->
-    <nav class="header-nav-menu hidden">
-      <button class="menu-close">Menu</button> <!-- Кнопка закриття меню -->
-      <ul>
-        <li><a href="#about">About me</a></li>
-        <li><a href="#benefits">Benefits</a></li>
-        <li><a href="#projects">Projects</a></li>
-        <li><a href="#faq">FAQ</a></li>
+     
+    <nav class="header-nav-menu">   
+      <button type="button" class="menu-close" aria-expanded="false" aria-controls="list-header">Menu</button> 
+       <ul class="nav-menu-list" id="list-header" aria-hidden="true">
+        <li class="nav-menu-item"><a class="menu-link-header" href="#about">About me</a></li>
+        <li class="nav-menu-item"><a class="menu-link-header" href="#benefits">Benefits</a></li>
+        <li class="nav-menu-item"><a class="menu-link-header" href="#projects">Projects</a></li>
+        <li class="nav-menu-item"><a class="menu-link-header" href="#faq">FAQ</a></li>
       </ul>
     </nav>
-    <!-- Кнопка замовлення проекту -->
     <a class="order-button" href="#work-together">Order the project</a>
   </div>
 </header>
+
+
+
